rename bottomnav & move into nav folder
